requirements.txt
ultralytics==8.2.0
streamlit==1.36.0
opencv-python-headless==4.8.1.78
pillow==10.0.0
numpy==1.26.4
